Split converters into their own files
