+++ lib/src/commands.dart
@@ -28,7 +28,9 @@ import 'context/autocomplete_context.dart';
 import 'context/base.dart';
 import 'context/chat_context.dart';
 import 'context/context_manager.dart';
+import 'converters/combine.dart';
 import 'converters/converter.dart';
+import 'converters/fallback.dart';
 import 'errors.dart';
 import 'mirror_utils/mirror_utils.dart';
 import 'options.dart';
