+++ lib/src/converters/converter.dart
@@ -20,10 +20,19 @@ import 'package:nyxx_interactions/nyxx_interactions.dart';
 import '../commands.dart';
 import '../context/autocomplete_context.dart';
 import '../context/base.dart';
-import '../context/chat_context.dart';
 import '../errors.dart';
 import '../mirror_utils/mirror_utils.dart';
 import '../util/view.dart';
+import 'built_in/attachment.dart';
+import 'built_in/bool.dart';
+import 'built_in/guild_channel.dart';
+import 'built_in/member.dart';
+import 'built_in/mentionable.dart';
+import 'built_in/number.dart';
+import 'built_in/role.dart';
+import 'built_in/snowflake.dart';
+import 'built_in/string.dart';
+import 'built_in/user.dart';
 
 /// Contains metadata and parsing capabilities for a given type.
 ///
@@ -140,996 +149,6 @@ class Converter<T> {
   @override
   String toString() => 'Converter<$T>';
 }
-
-/// A converter that extends the functionality of an existing converter, piping its output through
-/// another function.
-///
-/// This has the effect of allowing further processing of the output of a converter, for example to
-/// transform a [Snowflake] into a [IMember].
-///
-/// You might also be interested in:
-/// - [FallbackConverter], a converter that tries multiple converters successively.
-class CombineConverter<R, T> implements Converter<T> {
-  /// The converter used to parse the original input to the intermediate type.
-  final Converter<R> converter;
-
-  /// The function that transforms the intermediate type into the output type.
-  ///
-  /// As with normal converters, this function should not throw but can return `null` to indicate
-  /// parsing failure.
-  final FutureOr<T?> Function(R, IContextData) process;
-
-  @override
-  DartType<T> get output => DartType<T>();
-
-  final void Function(CommandOptionBuilder)? _customProcessOptionCallback;
-
-  @override
-  void Function(CommandOptionBuilder)? get processOptionCallback =>
-      _customProcessOptionCallback ?? converter.processOptionCallback;
-
-  final FutureOr<Iterable<ArgChoiceBuilder>?> Function(AutocompleteContext)? _autocompleteCallback;
-
-  @override
-  FutureOr<Iterable<ArgChoiceBuilder>?> Function(AutocompleteContext)? get autocompleteCallback =>
-      _autocompleteCallback ?? converter.autocompleteCallback;
-
-  @override
-  final FutureOr<MultiselectOptionBuilder> Function(T)? toMultiselectOption;
-
-  @override
-  final FutureOr<ButtonBuilder> Function(T)? toButton;
-
-  final Iterable<ArgChoiceBuilder>? _choices;
-  final CommandOptionType? _type;
-
-  /// Create a new [CombineConverter].
-  const CombineConverter(
-    this.converter,
-    this.process, {
-    Iterable<ArgChoiceBuilder>? choices,
-    CommandOptionType? type,
-    void Function(CommandOptionBuilder)? processOptionCallback,
-    FutureOr<Iterable<ArgChoiceBuilder>?> Function(AutocompleteContext)? autocompleteCallback,
-    this.toMultiselectOption,
-    this.toButton,
-  })  : _choices = choices,
-        _type = type,
-        _customProcessOptionCallback = processOptionCallback,
-        _autocompleteCallback = autocompleteCallback;
-
-  @override
-  Iterable<ArgChoiceBuilder>? get choices => _choices ?? converter.choices;
-
-  @override
-  CommandOptionType get type => _type ?? converter.type;
-
-  @override
-  FutureOr<T?> Function(StringView view, IContextData context) get convert =>
-      (view, context) async {
-        R? ret = await converter.convert(view, context);
-
-        if (ret != null) {
-          return await process(ret, context);
-        }
-        return null;
-      };
-
-  @override
-  String toString() => 'CombineConverter<$R, $T>[converter=$converter]';
-}
-
-/// A converter that successively tries a list of converters until one succeeds.
-///
-/// Given three converters *a*, *b* and *c*, a [FallbackConverter] will first try to convert the
-/// input using *a*, then, if *a* failed, using *b*, then, if *b* failed, using *c*. If all of *a*,
-/// *b* and *c* fail, then the [FallbackConverter] will also fail. If at least one of *a*, *b* or
-/// *c* succeed, the [FallbackConverter] will return the result of that conversion.
-///
-/// You might also be interested in:
-/// - [CombineConverter], for further processing the output of another converter.
-class FallbackConverter<T> implements Converter<T> {
-  /// The converters this [FallbackConverter] will attempt to use.
-  final Iterable<Converter<T>> converters;
-
-  @override
-  final void Function(CommandOptionBuilder)? processOptionCallback;
-
-  @override
-  final FutureOr<Iterable<ArgChoiceBuilder>?> Function(AutocompleteContext)? autocompleteCallback;
-
-  final Iterable<ArgChoiceBuilder>? _choices;
-  final CommandOptionType? _type;
-
-  final FutureOr<MultiselectOptionBuilder> Function(T)? _toMultiselectOption;
-
-  final FutureOr<ButtonBuilder> Function(T)? _toButton;
-
-  @override
-  DartType<T> get output => DartType<T>();
-
-  /// Create a new [FallbackConverter].
-  const FallbackConverter(
-    this.converters, {
-    Iterable<ArgChoiceBuilder>? choices,
-    CommandOptionType? type,
-    this.processOptionCallback,
-    this.autocompleteCallback,
-    FutureOr<MultiselectOptionBuilder> Function(T)? toMultiselectOption,
-    FutureOr<ButtonBuilder> Function(T)? toButton,
-  })  : _choices = choices,
-        _type = type,
-        _toMultiselectOption = toMultiselectOption,
-        _toButton = toButton;
-
-  @override
-  Iterable<ArgChoiceBuilder>? get choices {
-    if (_choices != null) {
-      return _choices;
-    }
-
-    List<ArgChoiceBuilder> allChoices = [];
-
-    for (final converter in converters) {
-      Iterable<ArgChoiceBuilder>? converterChoices = converter.choices;
-
-      if (converterChoices == null) {
-        return null;
-      }
-
-      for (final choice in converterChoices) {
-        ArgChoiceBuilder existing =
-            allChoices.singleWhere((element) => element.name == choice.name, orElse: () => choice);
-
-        if (existing.value != choice.value) {
-          return null;
-        } else if (identical(choice, existing)) {
-          allChoices.add(choice);
-        }
-      }
-    }
-
-    if (allChoices.isEmpty || allChoices.length > 25) {
-      return null;
-    }
-
-    return allChoices;
-  }
-
-  @override
-  CommandOptionType get type {
-    if (_type != null) {
-      return _type!;
-    }
-
-    Iterable<CommandOptionType> converterTypes = converters.map((converter) => converter.type);
-
-    if (converterTypes.every((element) => element == converterTypes.first)) {
-      return converterTypes.first;
-    }
-
-    return CommandOptionType.string;
-  }
-
-  @override
-  FutureOr<T?> Function(StringView view, IContextData context) get convert =>
-      (view, context) async {
-        StringView? used;
-        T? ret = await converters.fold(Future.value(null), (previousValue, element) async {
-          if (await previousValue != null) {
-            return await previousValue;
-          }
-
-          used = view.copy();
-          return await element.convert(used!, context);
-        });
-
-        if (used != null) {
-          view.history
-            ..clear()
-            ..addAll(used!.history);
-
-          view.index = used!.index;
-        }
-
-        return ret;
-      };
-
-  @override
-  FutureOr<MultiselectOptionBuilder> Function(T)? get toMultiselectOption {
-    if (_toMultiselectOption != null) {
-      return _toMultiselectOption;
-    }
-
-    for (final converter in converters) {
-      if (converter.toMultiselectOption != null) {
-        return converter.toMultiselectOption;
-      }
-    }
-
-    return null;
-  }
-
-  @override
-  FutureOr<ButtonBuilder> Function(T)? get toButton {
-    if (_toButton != null) {
-      return _toButton;
-    }
-
-    for (final converter in converters) {
-      if (converter.toButton != null) {
-        return converter.toButton;
-      }
-    }
-
-    return null;
-  }
-
-  @override
-  String toString() => 'FallbackConverter<$T>[converters=${List.of(converters)}]';
-}
-
-String? convertString(StringView view, IContextData context) => view.getQuotedWord();
-MultiselectOptionBuilder stringToMultiselectOption(String value) =>
-    MultiselectOptionBuilder(value, value);
-ButtonBuilder stringToButton(String value) => ButtonBuilder(value, '', ButtonStyle.primary);
-
-/// A [Converter] that converts input to a [String].
-///
-/// This converter returns the next space-separated word in the input, or, if the next word in the
-/// input is quoted, the next quoted section of the input.
-///
-/// This converter has a Discord Slash Command Argument Type of [CommandOptionType.string].
-const Converter<String> stringConverter = Converter<String>(
-  convertString,
-  type: CommandOptionType.string,
-  toMultiselectOption: stringToMultiselectOption,
-  toButton: stringToButton,
-);
-
-/// A converter that converts input to various types of numbers, possibly with a minimum or maximum
-/// value.
-///
-/// Note: this converter does not ensure that all values will be in the range `min..max`. [min] and
-/// [max] offer purely client-side validation and input from text commands is not validated beyond
-/// being a valid number.
-///
-/// You might also be interested in:
-/// - [IntConverter], for converting [int]s;
-/// - [DoubleConverter], for converting [double]s.
-class NumConverter<T extends num> extends Converter<T> {
-  /// The smallest value the user will be allowed to input in the Discord Client.
-  final T? min;
-
-  /// The biggest value the user will be allows to input in the Discord Client.
-  final T? max;
-
-  /// Create a new [NumConverter].
-  const NumConverter(
-    T? Function(StringView, IContextData) convert, {
-    required CommandOptionType type,
-    this.min,
-    this.max,
-  }) : super(convert, type: type);
-
-  @override
-  void Function(CommandOptionBuilder)? get processOptionCallback => (builder) {
-        builder.min = min;
-        builder.max = max;
-      };
-
-  @override
-  MultiselectOptionBuilder Function(T) get toMultiselectOption => (n) => MultiselectOptionBuilder(
-        n.toString(),
-        n.toString(),
-      );
-
-  @override
-  ButtonBuilder Function(T) get toButton => (n) => ButtonBuilder(
-        n.toString(),
-        '',
-        ButtonStyle.primary,
-      );
-}
-
-int? convertInt(StringView view, IContextData context) => int.tryParse(view.getQuotedWord());
-
-/// A converter that converts input to [int]s, possibly with a minimum or maximum value.
-///
-/// Note: this converter does not ensure that all values will be in the range `min..max`. [min] and
-/// [max] offer purely client-side validation and input from text commands is not validated beyond
-/// being a valid integer.
-///
-/// You might also be interested in:
-/// - [intConverter], the default [IntConverter].
-class IntConverter extends NumConverter<int> {
-  /// Create a new [IntConverter].
-  const IntConverter({
-    int? min,
-    int? max,
-  }) : super(
-          convertInt,
-          type: CommandOptionType.integer,
-          min: min,
-          max: max,
-        );
-}
-
-/// A [Converter] that converts input to an [int].
-///
-/// This converter attempts to parse the next word or quoted section of the input with [int.parse].
-///
-/// This converter has a Discord Slash Command Argument Type of [CommandOptionType.integer].
-const Converter<int> intConverter = IntConverter();
-
-double? convertDouble(StringView view, IContextData context) =>
-    double.tryParse(view.getQuotedWord());
-
-/// A converter that converts input to [double]s, possibly with a minimum or maximum value.
-///
-/// Note: this converter does not ensure that all values will be in the range `min..max`. [min] and
-/// [max] offer purely client-side validation and input from text commands is not validated beyond
-/// being a valid double.
-///
-/// You might also be interested in:
-/// - [doubleConverter], the default [DoubleConverter].
-class DoubleConverter extends NumConverter<double> {
-  /// Create a new [DoubleConverter].
-  const DoubleConverter({
-    double? min,
-    double? max,
-  }) : super(
-          convertDouble,
-          type: CommandOptionType.number,
-          min: min,
-          max: max,
-        );
-}
-
-/// A [Converter] that converts input to a [double].
-///
-/// This converter attempts to parse the next word or quoted section of the input with
-/// [double.parse].
-///
-/// This converter has a Discord Slash Command Argument Type of [CommandOptionType.number].
-const Converter<double> doubleConverter = DoubleConverter();
-
-bool? convertBool(StringView view, IContextData context) {
-  String word = view.getQuotedWord();
-
-  const Iterable<String> truthy = ['y', 'yes', '+', '1', 'true'];
-  const Iterable<String> falsy = ['n', 'no', '-', '0', 'false'];
-
-  const Iterable<String> valid = [...truthy, ...falsy];
-
-  if (valid.contains(word.toLowerCase())) {
-    return truthy.contains(word.toLowerCase());
-  }
-
-  return null;
-}
-
-MultiselectOptionBuilder boolToMultiselectOption(bool value) => MultiselectOptionBuilder(
-      value ? 'True' : 'False',
-      value.toString(),
-    );
-
-ButtonBuilder boolToButton(bool value) => ButtonBuilder(
-      value ? 'True' : 'False',
-      '',
-      ButtonStyle.primary,
-    );
-
-/// A [Converter] that converts input to a [bool].
-///
-/// This converter will parse the input to `true` if the next word or quoted section of the input is
-/// one of `y`, `yes`, `+`, `1` or `true`. This comparison is case-insensitive.
-/// This converter will parse the input to `false` if the next work or quoted section of the input
-/// is one of `n`, `no`, `-`, `0` or `false`. This comparison is case-insensitive.
-///
-/// If the input is not one of the aforementioned words, this converter will fail.
-///
-/// This converter has a Discord Slash Command Argument Type of [CommandOptionType.boolean].
-const Converter<bool> boolConverter = Converter<bool>(
-  convertBool,
-  type: CommandOptionType.boolean,
-  toMultiselectOption: boolToMultiselectOption,
-  toButton: boolToButton,
-);
-
-final RegExp _snowflakePattern = RegExp(r'^(?:<(?:@(?:!|&)?|#)([0-9]{15,20})>|([0-9]{15,20}))$');
-
-Snowflake? convertSnowflake(StringView view, IContextData context) {
-  String word = view.getQuotedWord();
-  if (!_snowflakePattern.hasMatch(word)) {
-    return null;
-  }
-
-  final RegExpMatch match = _snowflakePattern.firstMatch(word)!;
-
-  // 1st group will catch mentions, second will catch raw IDs
-  return Snowflake(match.group(1) ?? match.group(2));
-}
-
-MultiselectOptionBuilder snowflakeToMultiselectOption(Snowflake snowflake) =>
-    MultiselectOptionBuilder(
-      snowflake.toString(),
-      snowflake.toString(),
-    );
-
-ButtonBuilder snowflakeToButton(Snowflake snowflake) => ButtonBuilder(
-      snowflake.toString(),
-      '',
-      ButtonStyle.primary,
-    );
-
-/// A converter that converts input to a [Snowflake].
-///
-/// This converter will parse user mentions, member mentions, channel mentions or raw integers as
-/// snowflakes.
-const Converter<Snowflake> snowflakeConverter = Converter<Snowflake>(
-  convertSnowflake,
-  toMultiselectOption: snowflakeToMultiselectOption,
-  toButton: snowflakeToButton,
-);
-
-Future<IMember?> snowflakeToMember(Snowflake snowflake, IContextData context) async {
-  if (context.guild != null) {
-    IMember? cached = context.guild!.members[snowflake];
-    if (cached != null) {
-      return cached;
-    }
-
-    try {
-      return await context.guild!.fetchMember(snowflake);
-    } on IHttpResponseError {
-      return null;
-    }
-  }
-  return null;
-}
-
-Future<IMember?> convertMember(StringView view, IContextData context) async {
-  String word = view.getQuotedWord();
-
-  if (context.guild != null) {
-    Stream<IMember> named = context.guild!.searchMembersGateway(word, limit: 800000);
-
-    List<IMember> usernameExact = [];
-    List<IMember> nickExact = [];
-
-    List<IMember> usernameCaseInsensitive = [];
-    List<IMember> nickCaseInsensitive = [];
-
-    List<IMember> usernameStart = [];
-    List<IMember> nickStart = [];
-
-    await for (final member in named) {
-      IUser user = await member.user.getOrDownload();
-
-      if (user.username == word) {
-        usernameExact.add(member);
-      }
-      if (user.username.toLowerCase() == word.toLowerCase()) {
-        usernameCaseInsensitive.add(member);
-      }
-      if (user.username.toLowerCase().startsWith(word.toLowerCase())) {
-        usernameStart.add(member);
-      }
-
-      if (member.nickname != null) {
-        if (member.nickname! == word) {
-          nickExact.add(member);
-        }
-        if (member.nickname!.toLowerCase() == word.toLowerCase()) {
-          nickCaseInsensitive.add(member);
-        }
-        if (member.nickname!.toLowerCase().startsWith(word.toLowerCase())) {
-          nickStart.add(member);
-        }
-      }
-    }
-
-    for (final list in [
-      usernameExact,
-      nickExact,
-      usernameCaseInsensitive,
-      nickCaseInsensitive,
-      usernameStart,
-      nickStart
-    ]) {
-      if (list.length == 1) {
-        return list.first;
-      }
-    }
-  }
-  return null;
-}
-
-Future<MultiselectOptionBuilder> memberToMultiselectOption(IMember member) async {
-  IUser user = await member.user.getOrDownload();
-  String name = member.nickname ?? user.username;
-  String discriminator = user.formattedDiscriminator;
-
-  return MultiselectOptionBuilder(
-    '$name#$discriminator',
-    member.id.toString(),
-  );
-}
-
-Future<ButtonBuilder> memberToButton(IMember member) async {
-  IUser user = await member.user.getOrDownload();
-  String name = member.nickname ?? user.username;
-  String discriminator = user.formattedDiscriminator;
-
-  return ButtonBuilder(
-    '$name#$discriminator',
-    '',
-    ButtonStyle.primary,
-  );
-}
-
-/// A converter that converts input to an [IMember].
-///
-/// This will first attempt to parse the input to a snowflake which will then be converted to an
-/// [IMember]. If this fails, the member will be looked up by name.
-///
-/// This converter has a Discord Slash Command Argument Type of [CommandOptionType.user].
-const Converter<IMember> memberConverter = FallbackConverter<IMember>(
-  [
-    // Get member from mention or snowflake.
-    CombineConverter<Snowflake, IMember>(snowflakeConverter, snowflakeToMember),
-    // Get member by name or nickname
-    Converter<IMember>(convertMember),
-  ],
-  type: CommandOptionType.user,
-  toMultiselectOption: memberToMultiselectOption,
-  toButton: memberToButton,
-);
-
-Future<IUser?> snowflakeToUser(Snowflake snowflake, IContextData context) async {
-  IUser? cached = context.client.users[snowflake];
-  if (cached != null) {
-    return cached;
-  }
-
-  if (context.client is INyxxRest) {
-    try {
-      return await (context.client as INyxxRest).httpEndpoints.fetchUser(snowflake);
-    } on IHttpResponseError {
-      return null;
-    }
-  }
-
-  return null;
-}
-
-FutureOr<IUser?> memberToUser(IMember member, IContextData context) => member.user.getOrDownload();
-
-FutureOr<IUser?> convertUser(StringView view, IContextData context) {
-  String word = view.getWord();
-
-  if (context.channel.channelType == ChannelType.dm ||
-      context.channel.channelType == ChannelType.groupDm) {
-    List<IUser> exact = [];
-    List<IUser> caseInsensitive = [];
-    List<IUser> start = [];
-
-    for (final user in [
-      ...(context.channel as IDMChannel).participants,
-      if (context.client is INyxxRest) (context.client as INyxxRest).self,
-    ]) {
-      if (user.username == word) {
-        exact.add(user);
-      }
-
-      if (user.username.toLowerCase() == word.toLowerCase()) {
-        caseInsensitive.add(user);
-      }
-
-      if (user.username.toLowerCase().startsWith(word.toLowerCase())) {
-        start.add(user);
-      }
-
-      for (final list in [exact, caseInsensitive, start]) {
-        if (list.length == 1) {
-          return list.first;
-        }
-      }
-    }
-  }
-  return null;
-}
-
-MultiselectOptionBuilder userToMultiselectOption(IUser user) => MultiselectOptionBuilder(
-      '${user.username}#${user.formattedDiscriminator}',
-      user.id.toString(),
-    );
-
-ButtonBuilder userToButton(IUser user) => ButtonBuilder(
-      '${user.username}#${user.formattedDiscriminator}',
-      '',
-      ButtonStyle.primary,
-    );
-
-/// A converter that converts input to an [IUser].
-///
-/// This will first attempt to parse the input to a snowflake which will then be converted to an
-/// [IUser]. If this fails, the input will be parsed as an [IMember] which will then be converted to
-/// an [IUser]. If this fails, the user will be looked up by name.
-///
-/// This converter has a Discord Slash Command Argument Type of [CommandOptionType.user].
-const Converter<IUser> userConverter = FallbackConverter<IUser>(
-  [
-    CombineConverter<Snowflake, IUser>(snowflakeConverter, snowflakeToUser),
-    CombineConverter<IMember, IUser>(memberConverter, memberToUser),
-    Converter<IUser>(convertUser),
-  ],
-  type: CommandOptionType.user,
-  toMultiselectOption: userToMultiselectOption,
-  toButton: userToButton,
-);
-
-IGuildChannel? snowflakeToGuildChannel(Snowflake snowflake, IContextData context) {
-  if (context.guild != null) {
-    try {
-      return context.guild!.channels.firstWhere((channel) => channel.id == snowflake);
-    } on StateError {
-      return null;
-    }
-  }
-
-  return null;
-}
-
-IGuildChannel? convertGuildChannel(StringView view, IContextData context) {
-  if (context.guild != null) {
-    String word = view.getQuotedWord();
-
-    if (word.startsWith('#')) {
-      word = word.substring(1);
-    }
-
-    List<IGuildChannel> caseInsensitive = [];
-    List<IGuildChannel> partial = [];
-
-    for (final channel in context.guild!.channels) {
-      if (channel.name.toLowerCase() == word.toLowerCase()) {
-        caseInsensitive.add(channel);
-      }
-      if (channel.name.toLowerCase().startsWith(word.toLowerCase())) {
-        partial.add(channel);
-      }
-    }
-
-    for (final list in [caseInsensitive, partial]) {
-      if (list.length == 1) {
-        return list.first;
-      }
-    }
-  }
-
-  return null;
-}
-
-/// A converter that converts input to one or more types of [IGuildChannel]s.
-///
-/// This converter will only allow users to select channels of one of the types in [channelTypes],
-/// and then will further only accept channels of type `T`.
-///
-///
-/// Note: this converter does not ensure that all values will conform to [channelTypes].
-/// [channelTypes] offers purely client-side validation and input from text commands will not be
-/// validated beyond being assignable to `T`.
-///
-/// You might also be interested in:
-/// - [guildChannelConverter], a converter for all [IGuildChannel]s;
-/// - [textGuildChannelConverter], a converter for [ITextGuildChannel]s;
-/// - [voiceGuildChannelConverter], a converter for [IVoiceGuildChannel]s;
-/// - [categoryGuildChannelConverter], a converter for [ICategoryGuildChannel]s;
-/// - [stageVoiceChannelConverter], a converter for [IStageVoiceGuildChannel]s.
-class GuildChannelConverter<T extends IGuildChannel> implements Converter<T> {
-  /// The types of channels this converter allows users to select.
-  ///
-  /// If this is `null`, all channel types can be selected. Note that only channels which match both
-  /// these types *and* `T` will be parsed by this converter.
-  final List<ChannelType>? channelTypes;
-
-  final FallbackConverter<IGuildChannel> _internal = const FallbackConverter(
-    [
-      CombineConverter<Snowflake, IGuildChannel>(snowflakeConverter, snowflakeToGuildChannel),
-      Converter<IGuildChannel>(convertGuildChannel),
-    ],
-    type: CommandOptionType.channel,
-  );
-
-  /// Create a new [GuildChannelConverter].
-  const GuildChannelConverter(this.channelTypes);
-
-  @override
-  Iterable<ArgChoiceBuilder> get choices => [];
-
-  @override
-  FutureOr<T?> Function(StringView, IContextData) get convert => (view, context) async {
-        IGuildChannel? channel = await _internal.convert(view, context);
-
-        if (channel is T) {
-          return channel;
-        }
-
-        return null;
-      };
-
-  @override
-  void Function(CommandOptionBuilder) get processOptionCallback =>
-      (builder) => builder.channelTypes = channelTypes;
-
-  @override
-  FutureOr<Iterable<ArgChoiceBuilder>?> Function(AutocompleteContext)? get autocompleteCallback =>
-      null;
-
-  @override
-  DartType<T> get output => DartType<T>();
-
-  @override
-  CommandOptionType get type => CommandOptionType.channel;
-
-  @override
-  MultiselectOptionBuilder Function(T) get toMultiselectOption => (channel) {
-        MultiselectOptionBuilder builder = MultiselectOptionBuilder(
-          channel.name,
-          channel.id.toString(),
-        );
-
-        if (channel is ITextGuildChannel) {
-          builder.description = channel.topic;
-        }
-
-        return builder;
-      };
-
-  @override
-  ButtonBuilder Function(T) get toButton => (channel) => ButtonBuilder(
-        '#${channel.name}',
-        '',
-        ButtonStyle.primary,
-      );
-}
-
-/// A converter that converts input to an [IGuildChannel].
-///
-/// This will first attempt to parse the input as a [Snowflake] that will then be converted to an
-/// [IGuildChannel]. If this fails, the channel will be looked up by name in the current guild.
-///
-/// This converter has a Discord Slash Command argument type of [CommandOptionType.channel] and is
-/// set to accept all channel types.
-const GuildChannelConverter<IGuildChannel> guildChannelConverter = GuildChannelConverter(null);
-
-/// A converter that converts input to an [ITextGuildChannel].
-///
-/// This will first attempt to parse the input as a [Snowflake] that will then be converted to an
-/// [ITextGuildChannel]. If this fails, the channel will be looked up by name in the current guild.
-///
-/// This converter has a Discord Slash Command argument type of [CommandOptionType.channel] and is
-/// set to accept channels of type [ChannelType.text].
-const GuildChannelConverter<ITextGuildChannel> textGuildChannelConverter = GuildChannelConverter([
-  ChannelType.text,
-]);
-
-/// A converter that converts input to an [IVoiceGuildChannel].
-///
-/// This will first attempt to parse the input as a [Snowflake] that will then be converted to an
-/// [IVoiceGuildChannel]. If this fails, the channel will be looked up by name in the current guild.
-///
-/// This converter has a Discord Slash Command argument type of [CommandOptionType.channel] and is
-/// set to accept channels of type [ChannelType.voice].
-const GuildChannelConverter<IVoiceGuildChannel> voiceGuildChannelConverter = GuildChannelConverter([
-  ChannelType.voice,
-]);
-
-/// A converter that converts input to an [ICategoryGuildChannel].
-///
-/// This will first attempt to parse the input as a [Snowflake] that will then be converted to an
-/// [ICategoryGuildChannel]. If this fails, the channel will be looked up by name in the current
-/// guild.
-///
-/// This converter has a Discord Slash Command argument type of [CommandOptionType.channel] and it
-/// set to accept channels of type [ChannelType.category].
-const GuildChannelConverter<ICategoryGuildChannel> categoryGuildChannelConverter =
-    GuildChannelConverter([
-  ChannelType.category,
-]);
-
-/// A converter that converts input to an [IStageVoiceGuildChannel].
-///
-/// This will first attempt to parse the input as a [Snowflake] that will then be converted to an
-/// [IStageVoiceGuildChannel]. If this fails, the channel will be looked up by name in the current
-/// guild.
-///
-/// This converter has a Discord Slash Command argument type of [CommandOptionType.channel] and is
-/// set to accept channels of type [ChannelType.guildStage].
-const GuildChannelConverter<IStageVoiceGuildChannel> stageVoiceChannelConverter =
-    GuildChannelConverter([
-  ChannelType.guildStage,
-]);
-
-FutureOr<IRole?> snowflakeToRole(Snowflake snowflake, IContextData context) {
-  if (context.guild != null) {
-    IRole? cached = context.guild!.roles[snowflake];
-    if (cached != null) {
-      return cached;
-    }
-
-    try {
-      return context.guild!.fetchRoles().firstWhere((role) => role.id == snowflake);
-    } on StateError {
-      return null;
-    }
-  }
-
-  return null;
-}
-
-FutureOr<IRole?> convertRole(StringView view, IContextData context) async {
-  String word = view.getQuotedWord();
-  if (context.guild != null) {
-    Stream<IRole> roles = context.guild!.fetchRoles();
-
-    List<IRole> exact = [];
-    List<IRole> caseInsensitive = [];
-    List<IRole> partial = [];
-
-    await for (final role in roles) {
-      if (role.name == word) {
-        exact.add(role);
-      }
-      if (role.name.toLowerCase() == word.toLowerCase()) {
-        caseInsensitive.add(role);
-      }
-      if (role.name.toLowerCase().startsWith(word.toLowerCase())) {
-        partial.add(role);
-      }
-    }
-
-    for (final list in [exact, caseInsensitive, partial]) {
-      if (list.length == 1) {
-        return list.first;
-      }
-    }
-  }
-  return null;
-}
-
-MultiselectOptionBuilder roleToMultiselectOption(IRole role) {
-  MultiselectOptionBuilder builder = MultiselectOptionBuilder(
-    role.name,
-    role.id.toString(),
-  );
-
-  if (role.iconEmoji != null) {
-    builder.emoji = UnicodeEmoji(role.iconEmoji!);
-  }
-
-  return builder;
-}
-
-ButtonBuilder roleToButton(IRole role) => ButtonBuilder(
-      role.name,
-      '',
-      ButtonStyle.primary,
-    );
-
-/// A converter that converts input to an [IRole].
-///
-/// This will first attempt to parse the input as a snowflake that will then be converted to an
-/// [IRole]. If this fails, then the role will be looked up by name in the current guild.
-///
-/// This converter has a Discord Slash Command argument type of [CommandOptionType.role].
-const Converter<IRole> roleConverter = FallbackConverter<IRole>(
-  [
-    CombineConverter<Snowflake, IRole>(snowflakeConverter, snowflakeToRole),
-    Converter<IRole>(convertRole),
-  ],
-  type: CommandOptionType.role,
-  toMultiselectOption: roleToMultiselectOption,
-  toButton: roleToButton,
-);
-
-/// A converter that converts input to a [Mentionable].
-///
-/// This will first attempt to convert the input as a member, then as a role.
-///
-/// This converter has a Discord Slash Command argument type of [CommandOptionType.mentionable].
-const Converter<Mentionable> mentionableConverter = FallbackConverter(
-  [
-    memberConverter,
-    roleConverter,
-  ],
-  type: CommandOptionType.mentionable,
-);
-
-IAttachment? snowflakeToAttachment(Snowflake id, IContextData context) {
-  Iterable<IAttachment>? attachments;
-  if (context is InteractionChatContext) {
-    attachments = context.interaction.resolved?.attachments ?? [];
-  } else if (context is MessageChatContext) {
-    attachments = context.message.attachments;
-  }
-
-  if (attachments == null) {
-    return null;
-  }
-
-  try {
-    return attachments.singleWhere((attachment) => attachment.id == id);
-  } on StateError {
-    return null;
-  }
-}
-
-IAttachment? convertAttachment(StringView view, IContextData context) {
-  String fileName = view.getQuotedWord();
-
-  Iterable<IAttachment>? attachments;
-  if (context is InteractionChatContext) {
-    attachments = context.interaction.resolved?.attachments;
-  } else if (context is MessageChatContext) {
-    attachments = context.message.attachments;
-  }
-
-  if (attachments == null) {
-    return null;
-  }
-
-  Iterable<IAttachment> exactMatch = attachments.where(
-    (attachment) => attachment.filename == fileName,
-  );
-
-  Iterable<IAttachment> caseInsensitive = attachments.where(
-    (attachment) => attachment.filename.toLowerCase() == fileName.toLowerCase(),
-  );
-
-  Iterable<IAttachment> partialMatch = attachments.where(
-    (attachment) => attachment.filename.toLowerCase().startsWith(fileName.toLowerCase()),
-  );
-
-  for (final list in [exactMatch, caseInsensitive, partialMatch]) {
-    if (list.length == 1) {
-      return list.first;
-    }
-  }
-
-  return null;
-}
-
-MultiselectOptionBuilder attachmentToMultiselectOption(IAttachment attachment) =>
-    MultiselectOptionBuilder(
-      attachment.filename,
-      attachment.id.toString(),
-    );
-
-ButtonBuilder attachmentToButton(IAttachment attachment) => ButtonBuilder(
-      attachment.filename,
-      '',
-      ButtonStyle.primary,
-    );
-
-/// A converter that converts input to an [IAttachment].
-///
-/// This will first attempt to parse the input to a snowflake that will then be resolved as the ID
-/// of one of the attachments in the message or interaction. If this fails, then the attachment will
-/// be looked up by name.
-///
-/// This converter has a Discord Slash Command argument type of [CommandOptionType.attachment].
-const Converter<IAttachment> attachmentConverter = FallbackConverter(
-  [
-    CombineConverter<Snowflake, IAttachment>(snowflakeConverter, snowflakeToAttachment),
-    Converter(convertAttachment),
-  ],
-  type: CommandOptionType.attachment,
-  toMultiselectOption: attachmentToMultiselectOption,
-  toButton: attachmentToButton,
-);
 
 /// Apply a converter to an input and return the result.
 ///
